@Id @Generatedvalue added annotation
creating new package
creating Application class to run app
