+++ src/main/java/car/rental/model/BodyType.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 public enum BodyType {
