+++ src/main/java/car/rental/model/Branch.java
@@ -1,8 +1,7 @@
-package model;
+package car.rental.model;
 
 import lombok.Getter;
 import lombok.Setter;
-import net.bytebuddy.dynamic.loading.InjectionClassLoader;
 
 import javax.persistence.*;
 import java.util.List;
@@ -13,14 +12,16 @@ import java.util.List;
 @Table(name = "branch")
 public class Branch {
 
-    @Id
+    @Id     // The @Id annotation marks a field as a primary key field.
     @GeneratedValue(strategy = GenerationType.IDENTITY)
 
     @Column(name = "address")
-    private Address address;
+    private String address;
 
     @ OneToMany(mappedBy= "employee_list")
     private List<Employee> employeeList;
+
+    @OneToMany ( targetEntity = Car.class) //mappedBy = "car",
     @Column (name = "available_cars")
     private List<Car> availableCars;
 }
