+++ src/main/java/car/rental/model/Brand.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 public enum Brand {
