+++ src/main/java/car/rental/model/Color.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 public enum Color {
