+++ src/main/java/car/rental/model/Employee.java
@@ -1,10 +1,9 @@
-package model;
+package car.rental.model;
 
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.Table;
+import javax.persistence.*;
 
 @Getter
 @Setter
@@ -13,7 +12,9 @@ import javax.persistence.Table;
 
 public class Employee {
 
-    // Warning:(12, 1) Access can be package-private cos takie bylo jak komitowalam
+    @Id
+    @GeneratedValue (strategy = GenerationType.IDENTITY)
+
     private String name;
     private String surname;
     private String jobTitle;
