+++ src/main/java/car/rental/model/Model.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 public enum Model {
