+++ src/main/java/car/rental/model/OrderStatus.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 public enum OrderStatus {
