+++ src/main/java/car/rental/model/RentalOperation.java
@@ -1,10 +1,9 @@
-package model;
+package car.rental.model;
 
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.Entity;
-import javax.persistence.Table;
+import javax.persistence.*;
 import java.util.Date;
 
 @Getter
@@ -13,6 +12,10 @@ import java.util.Date;
 @Table (name = "rental_operation")
 
 public class RentalOperation {
+
+    @Id
+    @GeneratedValue (strategy = GenerationType.IDENTITY)
+
     private Employee employee;
     private Date FromDay;
     private Reservation reservation;
