+++ src/main/java/car/rental/model/Reservation.java
@@ -1,4 +1,4 @@
-package model;
+package car.rental.model;
 
 import lombok.Getter;
 import lombok.Setter;
